requirements.txt

google-genai>=1.0.0
httplib2>=0.20.4 